+++ src/test/Components/CreateImageWizard/steps/Packages/Packages.test.tsx
@@ -202,12 +202,12 @@ describe('Step Packages', () => {
     await verifyCancelButton(router);
   });
 
-  test('clicking Review and finish leads to Details', async () => {
+  test('clicking Review and finish leads to Review', async () => {
     await renderCreateMode();
     await goToPackagesStep();
     await clickReviewAndFinish();
     await screen.findByRole('heading', {
-      name: 'Details',
+      name: /Review/i,
     });
   });
